Remove FavoritesEmptyStateModel and FavoritesViewModel abstraction
