+++ DuckDuckGo/en.lproj/Localizable.strings
@@ -1787,9 +1787,6 @@ https://duckduckgo.com/mac";
 /* Title for the VPN Settings screen. */
 "network.protection.vpn.settings.title" = "VPN Settings";
 
-/* Header title of Favorites section */
-"new.tab.page.favorites.setion.header.title" = "Favorites";
-
 /* Information message about New Tab Page redesign */
 "new.tab.page.intro.message.body" = "Customize your Favorites and go-to features. Reorder things or hide them to keep it clean.";
