+++ DuckDuckGo/fr.lproj/Localizable.strings
@@ -1712,9 +1712,6 @@
 /* Title for the VPN Settings screen. */
 "network.protection.vpn.settings.title" = "Paramètres VPN";
 
-/* Header title of Favorites section */
-"new.tab.page.favorites.setion.header.title" = "Favoris";
-
 /* Information message about New Tab Page redesign */
 "new.tab.page.intro.message.body" = "Personnalisez vos favoris et vos fonctionnalités préférées. Réorganisez les différents éléments ou masquez-les pour que tout reste en ordre.";
